+++ code/server/globalv2loader.h
@@ -2,6 +2,6 @@
 
 #include "../qcommon/q_shared.h"
 qint
-xglobal_load_c(qchar *path);
+xglobal_load_c(const qchar *path);
 unsigned qchar
 xglobal_flags(qchar *ip);